+++ backend/canisters/user_index/impl/src/lifecycle/init.rs
@@ -1,5 +1,6 @@
 use crate::Data;
 use crate::lifecycle::{init_env, init_state};
+use crate::memory::get_stable_memory_map_memory;
 use canister_tracing_macros::trace;
 use ic_cdk::init;
 use tracing::info;
@@ -11,6 +12,7 @@ use utils::env::Environment;
 #[trace]
 fn init(args: Args) {
     canister_logger::init(args.test_mode);
+    stable_memory_map::init(get_stable_memory_map_memory());
     init_cycles_dispenser_client(args.cycles_dispenser_canister_id, args.test_mode);
 
     let env = init_env([0; 32], false);