+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -47,6 +47,8 @@ pub enum UserIndexEvent {
     AddCanisterToPool(CanisterId),
     ExternalAchievementAwarded(ExternalAchievementAwarded),
     SyncExistingUser(UserDetailsFull),
+    UserBlocked(UserId, UserId),
+    UserUnblocked(UserId, UserId),
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
@@ -124,7 +126,6 @@ pub struct BotRegistered {
     pub user_principal: Principal,
     pub name: String,
     pub commands: Vec<BotCommandDefinition>,
-    #[serde(default)]
     pub endpoint: String,
     pub autonomous_config: Option<AutonomousConfig>,
     pub permitted_install_location: Option<BotInstallationLocation>,
@@ -139,7 +140,6 @@ pub struct BotPublished {
 pub struct BotUpdated {
     pub bot_id: UserId,
     pub owner_id: UserId,
-    #[serde(default)]
     pub endpoint: String,
     pub definition: BotDefinition,
 }